# docs/swagger.yaml
basePath: /
definitions:
  dto.AdminResetPasswordRequest:
    properties:
      new_password:
        minLength: 4
        type: string
    required:
    - new_password
    type: object
...
  dto.UserBlockRequest:
    properties:
      is_blocked:
        description: pointer para permitir false (desbloquear) y detectar ausencia
          del campo
        type: boolean
    required:
    - is_blocked
    type: object
  dto.UserChangePasswordRequest:
    properties:
      current_password:
        minLength: 4
        type: string
      new_password:
        minLength: 4
        type: string
    required:
    - current_password
    - new_password
    type: object
  dto.UserCreateRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        minLength: 4
        type: string
      telp_number:
        maxLength: 20
        minLength: 7
        type: string
      username:
        maxLength: 20
        minLength: 4
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
...
  dto.UserMeUpdateRequest:
    properties:
      email:
        type: string
      image_url:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      telp_number:
        maxLength: 20
        minLength: 7
        type: string
      username:
        maxLength: 20
        minLength: 4
        type: string
    type: object
  dto.UserUpdateRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      telp_number:
        maxLength: 20
        minLength: 7
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta: {}
      status:
        type: boolean
    type: object
host: localhost:8888
info:
  contact:
    email: mddx56@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: API para gestionar motos
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Motos API
  version: 0.1.2
paths:
...
  /api/user:
    get:
      consumes:
      - application/json
      description: Get a list of all users with pagination and filtering
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort field
        in: query
        name: sort
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: order
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      - description: Relations to include (comma separated)
        in: query
        name: includes
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List all users
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User Create Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Create user
      tags:
      - user
  /api/user/{id}:
    get:
      consumes:
      - application/json
      description: Get user details by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update user details
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User Update Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - user
  /api/user/{id}/block:
    patch:
      consumes:
      - application/json
      description: Update user block status
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Block Request
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/dto.UserBlockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Block or unblock user
      tags:
      - user
  /api/user/{id}/devices:
    get:
      consumes:
      - application/json
      description: Get list of installed devices for a specific user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get installed devices
      tags:
      - user
  /api/user/{id}/reset-password:
    put:
      consumes:
      - application/json
      description: Admin resets a user's password by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Reset Password Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Reset user password (admin)
      tags:
      - user
  /api/user/{id}/status:
    patch:
      consumes:
      - application/json
      description: Toggle user status (true to false, false to true). Admin only.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Change user status (soft delete / restore)
      tags:
      - user
  /api/user/me:
    get:
      consumes:
      - application/json
      description: Get details of the currently logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update authenticated user data (without password). Validates unique
        email and username.
      parameters:
      - description: User Me Update Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UserMeUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update my profile
      tags:
      - user
  /api/user/me/password:
    put:
      consumes:
      - application/json
      description: Change authenticated user password (requires current_password)
      parameters:
      - description: Change Password Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Change my password
      tags:
      - user
  /api/user/simple:
    get:
      consumes:
      - application/json
      description: Get a list of users without pagination (only role user)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List simple users
      tags:
      - user
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"